Reagrouped everything

+++ SpaceMap.cpp
@@ -227,8 +227,8 @@ void SpaceMap::set_values()
         worlds[i].level_background.loadFromFile(address);
         worlds[i].level_sprite.setTexture(worlds[i].level_background);
         
-        float scaleX = (800) / worlds[i].level_background.getSize().x;
-        float scaleY = (600) / worlds[i].level_background.getSize().y;
+        float scaleX = static_cast<float>(800) / worlds[i].level_background.getSize().x;
+        float scaleY = static_cast<float>(600) / worlds[i].level_background.getSize().y;
         worlds[i].level_sprite.setScale(scaleX,scaleY);
     }
     rocketTexture.loadFromFile("./Images/ship.png");
